+++ tests/features/steps/crud_table.py
@@ -8,6 +8,7 @@ from __future__ import unicode_literals
 
 import wrappers
 from behave import when, then
+from textwrap import dedent
 
 
 @when('we create table')
@@ -63,7 +64,7 @@ def step_see_table_created(context):
     """
     Wait to see create table output.
     """
-    wrappers.expect_exact(context, 'CREATE TABLE', timeout=2)
+    wrappers.expect_pager(context, 'CREATE TABLE\r\n', timeout=2)
 
 
 @then('we see record inserted')
@@ -71,7 +72,7 @@ def step_see_record_inserted(context):
     """
     Wait to see insert output.
     """
-    wrappers.expect_exact(context, 'INSERT 0 1', timeout=2)
+    wrappers.expect_pager(context, 'INSERT 0 1\r\n', timeout=2)
 
 
 @then('we see record updated')
@@ -79,7 +80,7 @@ def step_see_record_updated(context):
     """
     Wait to see update output.
     """
-    wrappers.expect_exact(context, 'UPDATE 1', timeout=2)
+    wrappers.expect_pager(context, 'UPDATE 1\r\n', timeout=2)
 
 
 @then('we see data selected')
@@ -87,8 +88,17 @@ def step_see_data_selected(context):
     """
     Wait to see select output.
     """
-    wrappers.expect_exact(context, 'yyy', timeout=1)
-    wrappers.expect_exact(context, 'SELECT 1', timeout=1)
+    wrappers.expect_pager(
+        context,
+        dedent('''\
+            +-----+\r
+            | x   |\r
+            |-----|\r
+            | yyy |\r
+            +-----+\r
+            SELECT 1\r
+        '''),
+        timeout=1)
 
 
 @then('we see record deleted')
@@ -96,7 +106,7 @@ def step_see_data_deleted(context):
     """
     Wait to see delete output.
     """
-    wrappers.expect_exact(context, 'DELETE 1', timeout=2)
+    wrappers.expect_pager(context, 'DELETE 1\r\n', timeout=2)
 
 
 @then('we see table dropped')
@@ -104,4 +114,4 @@ def step_see_table_dropped(context):
     """
     Wait to see drop output.
     """
-    wrappers.expect_exact(context, 'DROP TABLE', timeout=2)
+    wrappers.expect_pager(context, 'DROP TABLE\r\n', timeout=2)