+++ tests/features/steps/named_queries.py
@@ -39,7 +39,7 @@ def step_see_named_query_saved(context):
     """
     Wait to see query saved.
     """
-    wrappers.expect_exact(context, 'Saved.', timeout=1)
+    wrappers.expect_pager(context, 'Saved.\r\n', timeout=1)
 
 
 @then('we see the named query executed')
@@ -56,4 +56,4 @@ def step_see_named_query_deleted(context):
     """
     Wait to see query deleted.
     """
-    wrappers.expect_exact(context, 'foo: Deleted', timeout=1)
+    wrappers.expect_pager(context, 'foo: Deleted\r\n', timeout=1)