+++ changelog.rst
@@ -17,6 +17,7 @@ Bug fixes:
 Internal changes:
 -----------------
 * Run pep8 checks in travis (Thanks: `Irina Truong`_).
+* Add pager wrapper for behave tests (Thanks: `Dick Marinus`_).
 
 1.5.1
 =====